Using cobra commands instead of several binaries
